refactor(vacs-client): split config into per-module config files
