+++ vacs-client/src/app/commands.rs
@@ -1,12 +1,12 @@
 use crate::app::state::AppState;
-use crate::app::{AppFolder, UpdateInfo, get_update, open_app_folder, open_fatal_error_dialog};
+use crate::app::{
+    AppFolder, ClientConfig, ClockMode, CplMode, FrontendCallConfig, FrontendClientPageSettings,
+    PersistedClientConfig, UpdateInfo, get_update, open_app_folder, open_fatal_error_dialog,
+};
 use crate::audio::manager::AudioManagerHandle;
 use crate::audio::source_type::SourceType;
 use crate::build::VersionInfo;
-use crate::config::{
-    AppConfig, CLIENT_SETTINGS_FILE_NAME, ClientConfig, ClockMode, CplMode, FrontendCallConfig,
-    FrontendClientPageSettings, Persistable, PersistedClientConfig,
-};
+use crate::config::{AppConfig, CLIENT_SETTINGS_FILE_NAME, Persistable};
 use crate::error::{Error, FrontendError};
 use crate::platform::Capabilities;
 use anyhow::Context;
