+++ vacs-client/src/app/state/signaling.rs
@@ -3,7 +3,7 @@ use crate::app::state::webrtc::{AppStateWebrtcExt, UnansweredCallGuard};
 use crate::app::state::{AppState, AppStateInner, sealed};
 use crate::audio::manager::AudioManagerHandle;
 use crate::audio::source_type::SourceType;
-use crate::config::{BackendEndpoint, WS_LOGIN_TIMEOUT};
+use crate::config::BackendEndpoint;
 use crate::error::{Error, FrontendError};
 use crate::signaling::auth::TauriTokenProvider;
 use serde::Serialize;
@@ -27,6 +27,7 @@ use vacs_signaling::protocol::ws::{client, server, shared};
 use vacs_signaling::transport::tokio::TokioTransport;
 
 const INCOMING_CALLS_LIMIT: usize = 5;
+const WS_LOGIN_TIMEOUT: Duration = Duration::from_secs(10);
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Default, Serialize)]
 #[serde(rename_all = "camelCase")]
