+++ vacs-client/src/keybinds/engine.rs
@@ -2,10 +2,9 @@ use crate::app::state::AppState;
 use crate::app::state::signaling::AppStateSignalingExt;
 use crate::app::state::webrtc::AppStateWebrtcExt;
 use crate::audio::manager::AudioManagerHandle;
-use crate::config::{CallMicMode, KeybindsConfig, TransmitConfig};
 use crate::error::Error;
 use crate::keybinds::runtime::{DynKeybindListener, KeybindListener, PlatformListener};
-use crate::keybinds::{KeyEvent, Keybind};
+use crate::keybinds::{CallMicMode, KeyEvent, Keybind, KeybindsConfig, TransmitConfig};
 use crate::radio::{RadioHandle, TransmissionState};
 use keyboard_types::{Code, KeyState};
 use parking_lot::RwLock;
