+++ vacs-client/src/signaling/commands.rs
@@ -2,11 +2,10 @@ use crate::app::state::http::HttpState;
 use crate::app::state::signaling::AppStateSignalingExt;
 use crate::app::state::webrtc::AppStateWebrtcExt;
 use crate::app::state::{AppState, AppStateInner};
+use crate::app::PersistedClientConfig;
 use crate::audio::manager::AudioManagerHandle;
 use crate::audio::source_type::SourceType;
-use crate::config::{
-    BackendEndpoint, CLIENT_SETTINGS_FILE_NAME, Persistable, PersistedClientConfig,
-};
+use crate::config::{BackendEndpoint, CLIENT_SETTINGS_FILE_NAME, Persistable};
 use crate::error::{Error, HandleUnauthorizedExt};
 use std::collections::HashSet;
 use tauri::{AppHandle, Manager, State};
